+++ src/main/java/org/generationcp/middleware/manager/api/GermplasmListManager.java
@@ -19,7 +19,12 @@ import org.generationcp.middleware.domain.gms.ListDataInfo;
 import org.generationcp.middleware.exceptions.MiddlewareQueryException;
 import org.generationcp.middleware.manager.Database;
 import org.generationcp.middleware.manager.Operation;
-import org.generationcp.middleware.pojos.*;
+import org.generationcp.middleware.pojos.GermplasmList;
+import org.generationcp.middleware.pojos.GermplasmListData;
+import org.generationcp.middleware.pojos.GermplasmListMetadata;
+import org.generationcp.middleware.pojos.ListDataProject;
+import org.generationcp.middleware.pojos.ListDataProperty;
+import org.generationcp.middleware.pojos.UserDefinedField;
 
 /**
  * This is the API for retrieving information about Germplasm Lists.
@@ -550,5 +555,5 @@ public interface GermplasmListManager {
 	 * Retrieves metadata (such as count of entries, list owner) in one go for all lists. This helps avoiding the need to query metadata in
 	 * a loop per list.
 	 */
-	Map<Long, GermplasmListMetadata> getAllGermplasmListMetadata();
+	Map<Integer, GermplasmListMetadata> getAllGermplasmListMetadata();
 }
