+++ src/test/java/org/generationcp/middleware/manager/GermplasmListManagerImplTest.java
@@ -22,7 +22,13 @@ import org.generationcp.middleware.domain.gms.ListDataInfo;
 import org.generationcp.middleware.exceptions.MiddlewareQueryException;
 import org.generationcp.middleware.manager.api.GermplasmDataManager;
 import org.generationcp.middleware.manager.api.GermplasmListManager;
-import org.generationcp.middleware.pojos.*;
+import org.generationcp.middleware.pojos.Germplasm;
+import org.generationcp.middleware.pojos.GermplasmList;
+import org.generationcp.middleware.pojos.GermplasmListData;
+import org.generationcp.middleware.pojos.GermplasmListMetadata;
+import org.generationcp.middleware.pojos.ListDataProject;
+import org.generationcp.middleware.pojos.Name;
+import org.generationcp.middleware.pojos.UserDefinedField;
 import org.generationcp.middleware.util.Util;
 import org.generationcp.middleware.utils.test.Debug;
 import org.hibernate.Query;
@@ -636,7 +642,7 @@ public class GermplasmListManagerImplTest extends IntegrationTestBase {
 	
 	@Test
 	public void testGetAllListMetadata() {
-		final Map<Long, GermplasmListMetadata> allGermplasmListMetadata = manager.getAllGermplasmListMetadata();
+		final Map<Integer, GermplasmListMetadata> allGermplasmListMetadata = manager.getAllGermplasmListMetadata();
 		Assert.assertNotNull("getAllGermplasmListMetadata() should never return null.", allGermplasmListMetadata);
 	}
